sorting, pagination, liking

+++ app/Http/Controllers/BlogController.php
@@ -7,6 +7,8 @@ use App\blogCategory;
 use App\blogPost;
 use App\Jobs\newBlogPostNotifications;
 use App\User;
+use Carbon\Carbon;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use Illuminate\Support\Str;
@@ -14,21 +16,148 @@ use Illuminate\Support\Facades\Auth;
 
 class BlogController extends Controller
 {
-    public function index(){
-        $posts = blogPost::all();
-        $categories = blogCategory::all();
+    public function index(Request $request){
 
-        $events = [
-            0 => [
-                'title' => 'Event Title1',
-                'start' => '2020-03-17T13:13:55.008',
-                'end'   => '2020-03-19T13:13:55.008',
-                'url'   => 'http://google.com/'
-            ]
-        ];
-        $events = json_encode($events);
+        if($request->ajax()){
+            $posts = $this->getSortResults($request);
+            $pagiStop = count($posts) == 5;
+            $html = view('partials.blog.posts')->withPosts($posts)->render();
 
-        return view('blogMainPage')->withEvents($events)->withPosts($posts)->withCats($categories);
+            return response()->json(['status' => 'success', 'html' => $html, 'stop' => $pagiStop], 200);
+        }else{
+            $categories = blogCategory::all();
+
+            $posts = $this->getSortResults($request);
+    
+            $events = [
+                0 => [
+                    'title' => 'Event Title1',
+                    'start' => '2020-03-17T13:13:55.008',
+                    'end'   => '2020-03-19T13:13:55.008',
+                    'url'   => 'http://google.com/'
+                ]
+            ];
+            $events = json_encode($events);
+    
+            return view('blogMainPage')->withEvents($events)->withPosts($posts)->withCats($categories);
+        }
+    }
+
+    private function getSortResults(Request $request)
+    {
+        $this->validateSortData($request);
+        return $this->getSortedPosts($request->all());
+    }
+
+    private function validateSortData(Request $data) : void
+    {
+        $data->validate([
+            'sortRange'     => ['string',Rule::in(['week','month','all'])],
+            'sortCrit'      => ['string',Rule::in(['date','likes'])],
+            'postCategory'  => ['exists:blog_categories,id','string'],
+            'postTags.*'    => ['string'],
+            'sortDir'       => ['string',Rule::in(['asc','desc'])],
+            'pagi'          => ['numeric','min:2']
+        ]);
+    }
+
+    private function getSortedPosts(array $criteria) : Collection
+    {
+
+        $posts = blogPost::select('*');
+        
+        if (isset($criteria['postCategory'])) {
+            $posts = $posts->where('category_id',$criteria['postCategory']);
+        }
+
+        if (isset($criteria['postTags'])) {
+            $posts = $posts->withAnyTag($criteria['postTags']);
+        }
+
+        if ( isset($criteria['sortCrit'])  && isset($criteria['sortDir'])) {
+            if ($criteria['sortCrit'] == "date") {
+                $posts = $posts->orderBy('created_at',$criteria['sortDir']);
+            }
+        }else{
+            $posts = $posts->orderBy('created_at','desc');
+        }
+        
+        if (isset($criteria['pagi'])) {
+            if (isset($criteria['sortCrit'])) {
+                if($criteria['sortCrit'] == 'likes'){
+                    if (isset($criteria['sortRange'])) {
+                        $range = $criteria['sortRange'];
+                    }else{
+                        $range = 'week';
+                    }
+                    switch ($range) {
+                        case 'week':
+                            $rangeTime = Carbon::now()->subWeek()->toDateTimeString();
+                            $posts = $posts->where('created_at','>',$rangeTime)->skip(5*(intVal($criteria['pagi']) - 1))->take(5)->get();
+                            break;
+                        case 'month':
+                            $rangeTime = Carbon::now()->subMonth()->toDateTimeString();
+                            $posts = $posts->where('created_at','>',$rangeTime)->skip(5*(intVal($criteria['pagi']) - 1))->take(5)->get();
+                            break;
+                        case 'all':
+                            $posts = $posts->skip(5*(intVal($criteria['pagi']) - 1))->take(5)->get();
+                            break;
+                    }
+                    
+                    if ($criteria['sortDir'] == "desc") {
+    
+                        $posts = $posts->sortByDesc(function($post,$key){
+                            return $post->likeCount;
+                        });
+                    }else{
+                        $posts = $posts->sortBy(function($post,$key){
+                            return $post->likeCount;
+                        });
+                    }
+                }
+            }else{
+                $posts = $posts->skip(5*(intVal($criteria['pagi']) - 1))->take(5)->get();
+            }
+        }else{
+            if (isset($criteria['sortCrit'])) {
+                if($criteria['sortCrit'] == 'likes'){
+                    if (isset($criteria['sortRange'])) {
+                        $range = $criteria['sortRange'];
+                    }else{
+                        $range = 'week';
+                    }
+                    switch ($range) {
+                        case 'week':
+                            $rangeTime = Carbon::now()->subWeek()->toDateTimeString();
+                            $posts = $posts->where('created_at','>',$rangeTime)->take(5)->get();
+                            break;
+                        case 'month':
+                            $rangeTime = Carbon::now()->subMonth()->toDateTimeString();
+                            $posts = $posts->where('created_at','>',$rangeTime)->take(5)->get();
+                            break;
+                        case 'all':
+                            $posts = $posts->take(5)->get();
+                            break;
+                    }
+                    
+                    if ($criteria['sortDir'] == "desc") {
+    
+                        $posts = $posts->sortByDesc(function($post,$key){
+                            return $post->likeCount;
+                        });
+                    }else{
+                        $posts = $posts->sortBy(function($post,$key){
+                            return $post->likeCount;
+                        });
+                    }
+                }
+            }else{
+                $posts = $posts->take(5)->get();
+            }
+        }
+        return $posts;
+
+
     }
 
     public function post(blogPost $blogPost)
@@ -162,9 +291,28 @@ class BlogController extends Controller
 
     public function deletePost(Request $request)
     {
-        $postId = $request->elementId;
+        if ($request->ajax()) {
+            $postId = $request->elementId;
+    
+            blogPost::find($postId)->delete();
+            return response()->json(['action' => "deletePostItem"], 200);
+        }
+    }
 
-        blogPost::find($postId)->delete();
-        return response()->json(['action' => "deletePostItem"], 200);
+    public function likePost(Request $request)
+    {
+        $request->validate([
+            'id'    => ['exists:blog_posts,id']
+        ]);
+
+        $post = blogPost::find($request->id);
+
+        if ($post->liked()) {
+            $post->unlike();
+        }else{
+            $post->like();
+        }
+
+        return response()->json(['status' => 'success'], 200);
     }
 }
